+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2 (2016-12-20)
+
+* Add BEACON output format
+* Implement diff command
+* Add arguments --sort and --no-cache
+
 ## 0.0.1 (2016-12-16)
 
 * First release at PyPI